Version 0.0.4
Some label changes and the fetch remote branches feature.
Added some variables to store the current buffer Git dir and the branch where the buffer was loaded.
Added autocmd's to fill the variables above.
Added an alert when the branch where the file is being saved is different from where it was loaded.
If the current buffer is not on the local dir, make a CD to the buffer dir when need to make a checkout or a fetch.
Changed to ask if we want to save a file loaded from one branch into another branch.