Version 0.0.8
Argh, another bug fixed. When trying to write to a file different from the current one, it was not saved. That happened because I was trying to save the file without it's name, presuming if we tell to save to another file with :w the name was correct.It was fixed using expand("<afile>"). Btw, there was another bug when trying to save to an existing file and need to use the bang (!) with write. It was fixed checking if a bang command was used and executing the bang version if needed.