Version 0.0.9
There were a bug when rebasing. On this situation HEAD just have a hash and not the information about the branches the plugin looks for. Thanks for Joachim Bargsten for warning me about this.
I made a double check on the HEAD file. If there is not the information the script wants, a message about the condition is returned. If something also "explodes" there, there is a try/catch block.
If detect this rebase/merge/bisect situation, the file will be always saved without the branch verification.